requirements.txt
requests
openai
python-dotenv 
exa-py